requirements.txt
asn1crypto==0.24.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
httplib2==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
macaroonbakery==1.1.3
Mako==1.0.7
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-dateutil==2.6.1
python-debian==0.1.32
python-telegram-bot==12.7
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.13.2
six==1.15.0
